Replace SID types
